delete confirmation implemented

+++ src/Pages/Dashboard/DoctorRow.js
@@ -4,34 +4,52 @@ import Swal from 'sweetalert2'
 const DoctorRow = ({ doctor, index, refetch }) => {
     const { name, specialty, img, email } = doctor
     const handleDelete = (email) => {
-        fetch(`http://localhost:5000/doctor/${email}`, {
-            method: "DELETE",
-            headers: {
-                "authorization": `Bearer ${localStorage.getItem("accessToken")}`
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`http://localhost:5000/doctor/${email}`, {
+                    method: "DELETE",
+                    headers: {
+                        "authorization": `Bearer ${localStorage.getItem("accessToken")}`
+                    }
+                })
+                    .then(res => res.json())
+                    .then(data => {
+                        if (data.deletedCount > 0) {
+                            Swal.fire({
+                                position: 'top-end',
+                                icon: 'success',
+                                title: 'Deleted Successfully',
+                                showConfirmButton: false,
+                                timer: 1500
+                            })
+                        }
+                        else {
+                            Swal.fire({
+                                position: 'top-end',
+                                icon: 'error',
+                                title: 'Deletion Failed',
+                                showConfirmButton: false,
+                                timer: 1500
+                            })
+                        }
+                        refetch();
+                    })
             }
         })
-            .then(res => res.json())
-            .then(data => {
-                if (data.deletedCount > 0) {
-                    Swal.fire({
-                        position: 'top-end',
-                        icon: 'success',
-                        title: 'Deleted Successfully',
-                        showConfirmButton: false,
-                        timer: 1500
-                    })
-                }
-                else {
-                    Swal.fire({
-                        position: 'top-end',
-                        icon: 'error',
-                        title: 'Deletion Failed',
-                        showConfirmButton: false,
-                        timer: 1500
-                    })
-                }
-                refetch();
-            })
+
+
+
+
+
+
 
     }
